# src/main/java/org/openapitools/api/LoadApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Error;
import org.openapitools.model.LoadRequest;
import org.openapitools.model.LoadResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-04-25T23:46:54.159924+05:30[Asia/Kolkata]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "load", description = "the load API")
public interface LoadApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * PUT /load/{messageId} : Adds funds to a user&#39;s account.
     *
     * @param messageId Unique identifier for a message (required)
     * @param loadRequest An load request message that needs to be decisioned. This balance will be added to a user&#39;s balance. (optional)
     * @return The result of an load (status code 201)
     *         or Server Error response (status code 200)
     */
    @Operation(
        operationId = "loadMessageIdPut",
        summary = "Adds funds to a user's account.",
        responses = {
            @ApiResponse(responseCode = "201", description = "The result of an load", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LoadResponse.class))
            }),
            @ApiResponse(responseCode = "default", description = "Server Error response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/load/{messageId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<LoadResponse> loadMessageIdPut(
        @Parameter(name = "messageId", description = "Unique identifier for a message", required = true, in = ParameterIn.PATH) @PathVariable("messageId") String messageId,
        @Parameter(name = "LoadRequest", description = "An load request message that needs to be decisioned. This balance will be added to a user's balance.") @Valid @RequestBody(required = false) LoadRequest loadRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"balance\" : { \"amount\" : \"amount\", \"currency\" : \"currency\" }, \"messageId\" : \"messageId\", \"userId\" : \"userId\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}